requirements.txt
asgiref==3.7.2
Django==5.0.3
django-crispy-forms==2.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
